requirements.txt
aiogram>=3.7.0
httpx>=0.27.0
PyYAML>=6.0.1
pydantic>=2.11.0
